file-searcher/requirements.txt
# GUI框架
PyQt5>=5.15.0

# 全文检索
whoosh>=2.7.4

# 文件解析
python-docx>=0.8.11
openpyxl>=3.1.2
pdfplumber>=0.10.3
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=4.9.3

# 系统工具
psutil>=5.9.6

# 其他
python-dateutil>=2.8.2
